FULL CODE/requirements.txt
opencv-python
numpy
torch
torchvision
ultralytics
transformers
sentence-transformers
faiss-cpu
flask
fastapi
uvicorn
python-multipart
pyyaml
pyttsx3
pillow
torch-geometric
twilio
requests
google-generativeai
soundfile
